core: FIX: autocompile / hotcompile wasn't refreshing classpath resources properly
Using HotReloadingClassloader wasn't useful since a new classloader is used after each compilation, and HotReloadingClassloader was customized to use CompilationManager destination dir for class loading only, not resource loading. In the end the simplest and easiest is to use a URLClassloader targetted on compilation destination, and renewed after each compilation.