server: check server socket can be opened before starting the server
This makes the server start actually fail when it isn't able top open the port, which is easier to diagnose especially during tests.

Add better error reporting for "address already in use" which is one of the most common case during development, and the message is not always clearly understood, especially by beginners.