core: perf improvement: remove unnecessary list copy
This copy was previously necessary because MongoIterator implements Iterable but can't return a new iterator on each call, which makes some code relying on that broken. But with properly configured Jackons writer we don't fall into this problem anymore, and can remove that copy. I needed a check for Iterator would be better.