factory - introduce priority at engine level [BREAKING]
allow to set priority for each provided component in a Module
store priority in the named component, making it possible to use that information
to sort components of different types per priority

[BREAKING]

This is introduce some changes at API level of some factory related
classes but they should be backward compatible most of the time, and
shouldn't be used too frequently. You may notice a breaking change if
you implemented the MachineEngine interface directly.

Moreover, this introduce a change in the order of components, using the
component name as second key to sort components. So while the order was
not fully deterministic when 2 machines had the same priority before,
now it is deterministic and this may impact your code.