Adds "named" attribute to Alternative annotation
This attribute permits to force the name which will be used for the
Alternative component. It's optional, so if it is not defined the default
behavior will be used (SimpleName of the referenced class, if the referenced
class does not specify a name, or the specified name of the registered class
if it defines a name using @Named annotation).