Allow to plug other cache mechanisms for RestxSession #141 [breaking]
[breaking] RestxSession.Definition.Entry is now an interface.
Use DefaultSessionDefinitionEntry to get similar behavior if you were instanciating
RestxSession.Definition.Entry before.

[breaking] RestxSession#cleanUpCaches has been removed
It was misleading, was not invalidating the cache and very implementation dependent.
2 invalidate methods have been added to replace it.

---
this commit review the responsibilities introduced in previous commit, to make the cache
pluggable without having to rewrite all Definition.Entry. Now the Entry knows how to load,
and the EntryCacheManager knows how to cache, so there is a better separation of responsibilities.