introduces restx.cold.classes property
This property contains a string with all FQCN of cold classes, separated by ':'.

This property might be specified using System.properties like that:
```
System.setProperty("restx.cold.classes", "foo.ColdClass1:bar.ColdClass2");
```
It meant that `foo.ColdClass1` and `bar.ColdClass2` must be ignored by hot-reload
mechanism.

As cold classes are supposed to be used by components, they must be loaded with the same classloader
as the one used for components, that's why the classloader used to initialize the factory's components
is kept, and re-used to transform cold classes FQCN into real classes.