Reworked war overlay to depends on ui in a profile
See #187 (and also #180)

War overlay plugin was depending on -ui.zip artifact, replaced by a profile to prevent build errors when, the ui artifact is nowhere to be found.
If you want to build the "final" artifact, you will need to run the package command with the "release" profile enabled :

    mvn package -Prelease

This will create a war file ending in -full.war containing your server logic as well as your ui.
This generated file exists to prevent name clashing between artifact with and without ui logic.