fix: exception when watcher services are closed
When watcher are closed and there is still a running thread waiting for some event,
an java.nio.file.ClosedWatchServiceException is thrown. According to the
javadoc (https://docs.oracle.com/javase/7/docs/api/java/nio/file/WatchService.html#take%28%29)
this is the expected behavior, so we don't need to propagate this exception outside the thread.