module are able to produces generics
AnnotationProcessor for @Module have been modified in order to be able to provides
components having a parameterized type.

Also in modules we will be able to provides generics conditional, alternatives will not be possible
though, as annotations does not allow a Type to be used as a parameter (the idea would have been
to replace the type of "to" from Class to Type).

wip: check during compilation that the generic type of the component is not a "reserved" one,
like List, Iterable, that are used to inject all components of a specified type.