params - allowing to handle complex types as query parameters
Complex types are deserialized through Jackson after creating a hierarchical map from query parameters
The only complexity is to identify if a node in this hierarchical map should be an array of raw types or a single value.
I introduced a ParamDef.SpecialParamHandling class aimed at describing this.

Also, note that iterable of complex types are not handled as it would require to name parameters like this : foo[1].bar
and I guess we all want to avoid this