feat: upgrade jackson, jongo and mongo driver [breaking]
upgrade to latest versions of jackson jongo and mongo driver

latest jongo (1.4.0) has breaking changes compared to 1.1 version used before:
if you use `@Id String myproperty` in your Entities, the id generated
by mongo was populated in the property in Jongo 1.1. This is not the
case anymore, for that you need to add the `@MongoObjectId` annotation.
It is also recommended to replace @Id by @MongoId.
(but then it will be named _id in mongo but not in rest resources.)